Add models file and move the structs
